add profile for spring data jpa services | correct AbstractMapService id type for PetTypeServiceMap

+++ pet-clinic-data/src/main/java/org/learning/spring/guruspringpetclinic/repositories/SpecialityRepository.java
@@ -3,5 +3,5 @@ package org.learning.spring.guruspringpetclinic.repositories;
 import org.learning.spring.guruspringpetclinic.model.Speciality;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SpecialityRepository extends CrudRepository<Speciality, Long> {
+public interface SpecialityRepository extends CrudRepository<Speciality, Integer> {
 }
